+++ R/landscape_summary.R
@@ -8,8 +8,11 @@
 #' @param name A string of the element to summarize: "forest", "soil" or "state".
 #' @param summaryFunction A function that accepts objects of class \code{\link{forest}}, \code{soil} or model input objects, respectively.
 #' @param ... Additional arguments to the summary function.
+#' @param unlist Logical flag to try converting the summaries into different columns
 #' 
 #' @returns An object of class \code{\link{sf}} containing the calculated statistics. 
+#' If \code{unlist = FALSE} column 'summary' is a list with summaries for each element. 
+#' If \code{unlist = TRUE} different columns are returned instead, one per variable given in the summary function.
 #' 
 #' @author Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF.
 #' 
@@ -28,7 +31,7 @@
 #' # Apply forest summary function
 #' landscape_summary(y, "forest", summary.forest, SpParamsMED)
 #'  
-landscape_summary<-function(object, name, summaryFunction, ...) {
+landscape_summary<-function(object, name, summaryFunction, ..., unlist = FALSE) {
   if(!inherits(object, "sf")) stop("'object' has to be an object of class 'sf'.")
   name = match.arg(name, c("forest", "soil", "state"))
   l = object[[name]]
@@ -39,6 +42,15 @@ landscape_summary<-function(object, name, summaryFunction, ...) {
     if(!l_isnull[i]) sm[[i]] = do.call(summaryFunction, args=list(object=l[[i]],...))
   }
   res = sf::st_sf(geometry=sf::st_geometry(y))
-  res$summary = sm
+  if(!unlist) {
+    res$summary = sm
+  } else {
+    d1 = unlist(sm[[1]])
+    nms = names(d1)
+    for(nm in nms) res[[nm]] = NA
+    for(i in 1:length(sm)) {
+      res[i, nms] = unlist(sm[[i]])
+    }
+  }
   return(sf::st_as_sf(tibble::as_tibble(res)))
 }